@ core: make the real-time DSP pipeline pure f32 (no f64 sample widening)
Sample is f32, but the per-sample path kept widening amplitudes back to f64
(gather sums, biquad state/coeffs, stream interp, mix gains, slewer, delay).
Purge all of it so audio amplitudes stay f32 end to end.

Resampler: now contains zero f64. Kernel tables build in f32 with Kahan-summed
integral (keeps accuracy); lookups, gathers (scalar + SIMD phasor), positions,
cutoffs, and index math are all f32. Long-note pitch drift is handled by keeping
the phase bounded, not by f64 precision:
  - voices fold the loop by subtracting the period (not modulo), so sample_t
    stays inside the loop body and f32-precise forever;
  - the StreamResampler splits its open-ended position into an exact i64 integer
    part + an f32 fraction in [0,1), so the kernel never sees a large float and
    the stream cannot drift.

BiquadFilter: f32 state + coefficients (computed in f64, narrowed once); the
per-sample transform is pure f32. Analysis (response/poles/zeros) stays f64.
Added an f32-stability test at the bass crossover + PSG-comp cutoffs.

Slewer/DelayLine/instrument volume+gain/synth pan+volume are f32; device control
values (dB-domain envelopes, pan law) stay f64 and are narrowed once at the
device -> synth boundary.

Loosened a couple of test tolerances that were tighter than f32 epsilon.

CLAUDE.md: forbid launching subagents.
@

+++ crates/optime-core/src/dsp/resample/gather.rs
@@ -3,7 +3,7 @@
 
 #[cfg(not(feature = "simd"))]
 use super::kernels::sinc_at;
-use super::kernels::{sinc_int_at, win_at, Kernels};
+use super::kernels::{Kernels, sinc_int_at, win_at};
 
 /// Scalar BLEP gather of the boxcar-integrated windowed kernel: tap `j` weighs its source-sample
 /// bin `[pos−k−1, pos−k]` (where `d_hi = d0 − j = pos − k`) by the band-limited step rise across
@@ -15,19 +15,19 @@ use super::kernels::{sinc_int_at, win_at, Kernels};
 pub(crate) fn gather_step(
     k: &Kernels,
     src: &[f32],
-    d0: f64,
-    sinc_idx_step: f64,
-    win_idx_step: f64,
-) -> (f64, f64) {
-    let mut out = 0.0;
-    let mut wsum = 0.0;
+    d0: f32,
+    sinc_idx_step: f32,
+    win_idx_step: f32,
+) -> (f32, f32) {
+    let mut out = 0.0f32;
+    let mut wsum = 0.0f32;
     let mut si_hi = sinc_int_at(k, sinc_idx_step * d0);
     let mut lo_idx = sinc_idx_step * (d0 - 1.0); // S index of the bin's lower edge
     let mut mid_idx = win_idx_step * (d0 - 0.5); // window index of the bin centre (signed)
     for &s in src {
         let si_lo = sinc_int_at(k, lo_idx);
         let w = win_at(k, mid_idx.abs()) * (si_hi - si_lo);
-        out += f64::from(s) * w;
+        out += s * w;
         wsum += w;
         si_hi = si_lo;
         lo_idx -= sinc_idx_step;
@@ -45,13 +45,13 @@ pub(crate) fn gather_step(
 pub(crate) fn gather_impulse(
     k: &Kernels,
     src: &[f32],
-    d0: f64,
+    d0: f32,
     mid_j: usize,
-    sinc_idx_step: f64,
-    win_idx_step: f64,
-) -> (f64, f64) {
-    let mut out = 0.0;
-    let mut wsum = 0.0;
+    sinc_idx_step: f32,
+    win_idx_step: f32,
+) -> (f32, f32) {
+    let mut out = 0.0f32;
+    let mut wsum = 0.0f32;
     let (right, left) = src.split_at(mid_j + 1);
 
     // Right run: descending |d| = d0 − j.
@@ -59,18 +59,18 @@ pub(crate) fn gather_impulse(
     let mut win_idx = d0 * win_idx_step;
     for &s in right {
         let w = sinc_at(k, sinc_idx) * win_at(k, win_idx);
-        out += f64::from(s) * w;
+        out += s * w;
         wsum += w;
         sinc_idx -= sinc_idx_step;
         win_idx -= win_idx_step;
     }
     // Left run: ascending |d| = j − d0.
-    let d_left = mid_j as f64 + 1.0 - d0;
+    let d_left = mid_j as f32 + 1.0 - d0;
     let mut sinc_idx = d_left * sinc_idx_step;
     let mut win_idx = d_left * win_idx_step;
     for &s in left {
         let w = sinc_at(k, sinc_idx) * win_at(k, win_idx);
-        out += f64::from(s) * w;
+        out += s * w;
         wsum += w;
         sinc_idx += sinc_idx_step;
         win_idx += win_idx_step;
